Revert version misalib

+++ mesalib/src/util/hash_table.c
@@ -47,31 +47,8 @@
 #include "hash_table.h"
 #include "ralloc.h"
 #include "macros.h"
-#include "u_memory.h"
+#include "main/hash.h"
 #include "fast_urem_by_const.h"
-#include "util/u_memory.h"
-
-#define XXH_INLINE_ALL
-#include "xxhash.h"
-
-/**
- * Magic number that gets stored outside of the struct hash_table.
- *
- * The hash table needs a particular pointer to be the marker for a key that
- * was deleted from the table, along with NULL for the "never allocated in the
- * table" marker.  Legacy GL allows any GLuint to be used as a GL object name,
- * and we use a 1:1 mapping from GLuints to key pointers, so we need to be
- * able to track a GLuint that happens to match the deleted key outside of
- * struct hash_table.  We tell the hash table to use "1" as the deleted key
- * value, so that we test the deleted-key-in-the-table path as best we can.
- */
-#define DELETED_KEY_VALUE 1
-
-static inline void *
-uint_key(unsigned id)
-{
-   return (void *)(uintptr_t) id;
-}
 
 static const uint32_t deleted_key_value;
 
@@ -121,7 +98,7 @@ static const struct {
    ENTRY(2147483648ul, 2362232233ul, 2362232231ul )
 };
 
-ASSERTED static inline bool
+static inline bool
 key_pointer_is_reserved(const struct hash_table *ht, const void *key)
 {
    return key == NULL || key == ht->deleted_key;
@@ -191,26 +168,6 @@ _mesa_hash_table_create(void *mem_ctx,
    return ht;
 }
 
-static uint32_t
-key_u32_hash(const void *key)
-{
-   uint32_t u = (uint32_t)(uintptr_t)key;
-   return _mesa_hash_uint(&u);
-}
-
-static bool
-key_u32_equals(const void *a, const void *b)
-{
-   return (uint32_t)(uintptr_t)a == (uint32_t)(uintptr_t)b;
-}
-
-/* key == 0 and key == deleted_key are not allowed */
-struct hash_table *
-_mesa_hash_table_create_u32_keys(void *mem_ctx)
-{
-   return _mesa_hash_table_create(mem_ctx, key_u32_hash, key_u32_equals);
-}
-
 struct hash_table *
 _mesa_hash_table_clone(struct hash_table *src, void *dst_mem_ctx)
 {
@@ -254,13 +211,6 @@ _mesa_hash_table_destroy(struct hash_table *ht,
    ralloc_free(ht);
 }
 
-static void
-hash_table_clear_fast(struct hash_table *ht)
-{
-   memset(ht->table, 0, sizeof(struct hash_entry) * hash_sizes[ht->size_index].size);
-   ht->entries = ht->deleted_entries = 0;
-}
-
 /**
  * Deletes all entries of the given hash table without deleting the table
  * itself or changing its structure.
@@ -271,22 +221,20 @@ void
 _mesa_hash_table_clear(struct hash_table *ht,
                        void (*delete_function)(struct hash_entry *entry))
 {
-   if (!ht)
-      return;
-
    struct hash_entry *entry;
 
-   if (delete_function) {
-      for (entry = ht->table; entry != ht->table + ht->size; entry++) {
-         if (entry_is_present(ht, entry))
-            delete_function(entry);
+   for (entry = ht->table; entry != ht->table + ht->size; entry++) {
+      if (entry->key == NULL)
+         continue;
 
-         entry->key = NULL;
-      }
-      ht->entries = 0;
-      ht->deleted_entries = 0;
-   } else
-      hash_table_clear_fast(ht);
+      if (delete_function != NULL && entry->key != ht->deleted_key)
+         delete_function(entry);
+
+      entry->key = NULL;
+   }
+
+   ht->entries = 0;
+   ht->deleted_entries = 0;
 }
 
 /** Sets the value of the key pointer used for deleted entries in the table.
@@ -306,7 +254,7 @@ _mesa_hash_table_set_deleted_key(struct hash_table *ht, const void *deleted_key)
 }
 
 static struct hash_entry *
-hash_table_search(const struct hash_table *ht, uint32_t hash, const void *key)
+hash_table_search(struct hash_table *ht, uint32_t hash, const void *key)
 {
    assert(!key_pointer_is_reserved(ht, key));
 
@@ -342,7 +290,7 @@ hash_table_search(const struct hash_table *ht, uint32_t hash, const void *key)
  * modified by the user.
  */
 struct hash_entry *
-_mesa_hash_table_search(const struct hash_table *ht, const void *key)
+_mesa_hash_table_search(struct hash_table *ht, const void *key)
 {
    assert(ht->key_hash_function);
    return hash_table_search(ht, ht->key_hash_function(key), key);
@@ -391,12 +339,6 @@ _mesa_hash_table_rehash(struct hash_table *ht, unsigned new_size_index)
    struct hash_table old_ht;
    struct hash_entry *table;
 
-   if (ht->size_index == new_size_index && ht->deleted_entries == ht->max_entries) {
-      hash_table_clear_fast(ht);
-      assert(!ht->entries);
-      return;
-   }
-
    if (new_size_index >= ARRAY_SIZE(hash_sizes))
       return;
 
@@ -427,7 +369,8 @@ _mesa_hash_table_rehash(struct hash_table *ht, unsigned new_size_index)
 }
 
 static struct hash_entry *
-hash_table_get_entry(struct hash_table *ht, uint32_t hash, const void *key)
+hash_table_insert(struct hash_table *ht, uint32_t hash,
+                  const void *key, void *data)
 {
    struct hash_entry *available_entry = NULL;
 
@@ -468,8 +411,11 @@ hash_table_get_entry(struct hash_table *ht, uint32_t hash, const void *key)
        */
       if (!entry_is_deleted(ht, entry) &&
           entry->hash == hash &&
-          ht->key_equals_function(key, entry->key))
+          ht->key_equals_function(key, entry->key)) {
+         entry->key = key;
+         entry->data = data;
          return entry;
+      }
 
       hash_address += double_hash;
       if (hash_address >= size)
@@ -480,6 +426,8 @@ hash_table_get_entry(struct hash_table *ht, uint32_t hash, const void *key)
       if (entry_is_deleted(ht, available_entry))
          ht->deleted_entries--;
       available_entry->hash = hash;
+      available_entry->key = key;
+      available_entry->data = data;
       ht->entries++;
       return available_entry;
    }
@@ -488,20 +436,6 @@ hash_table_get_entry(struct hash_table *ht, uint32_t hash, const void *key)
     * application could ignore this result.
     */
    return NULL;
-}
-
-static struct hash_entry *
-hash_table_insert(struct hash_table *ht, uint32_t hash,
-                  const void *key, void *data)
-{
-   struct hash_entry *entry = hash_table_get_entry(ht, hash, key);
-
-   if (entry) {
-      entry->key = key;
-      entry->data = data;
-   }
-
-   return entry;
 }
 
 /**
@@ -550,27 +484,6 @@ void _mesa_hash_table_remove_key(struct hash_table *ht,
                                  const void *key)
 {
    _mesa_hash_table_remove(ht, _mesa_hash_table_search(ht, key));
-}
-
-/**
- * This function is an iterator over the hash_table when no deleted entries are present.
- *
- * Pass in NULL for the first entry, as in the start of a for loop.
- */
-struct hash_entry *
-_mesa_hash_table_next_entry_unsafe(const struct hash_table *ht, struct hash_entry *entry)
-{
-   assert(!ht->deleted_entries);
-   if (!ht->entries)
-      return NULL;
-   if (entry == NULL)
-      entry = ht->table;
-   else
-      entry = entry + 1;
-   if (entry != ht->table + ht->size)
-      return entry->key ? entry : _mesa_hash_table_next_entry_unsafe(ht, entry);
-
-   return NULL;
 }
 
 /**
@@ -633,80 +546,35 @@ _mesa_hash_table_random_entry(struct hash_table *ht,
 }
 
 
+/**
+ * Quick FNV-1a hash implementation based on:
+ * http://www.isthe.com/chongo/tech/comp/fnv/
+ *
+ * FNV-1a is not be the best hash out there -- Jenkins's lookup3 is supposed
+ * to be quite good, and it probably beats FNV.  But FNV has the advantage
+ * that it involves almost no code.  For an improvement on both, see Paul
+ * Hsieh's http://www.azillionmonkeys.com/qed/hash.html
+ */
 uint32_t
 _mesa_hash_data(const void *data, size_t size)
 {
-   return XXH32(data, size, 0);
-}
-
-uint32_t
-_mesa_hash_data_with_seed(const void *data, size_t size, uint32_t seed)
-{
-   return XXH32(data, size, seed);
-}
-
-uint32_t
-_mesa_hash_int(const void *key)
-{
-   return XXH32(key, sizeof(int), 0);
-}
-
-uint32_t
-_mesa_hash_uint(const void *key)
-{
-   return XXH32(key, sizeof(unsigned), 0);
-}
-
-uint32_t
-_mesa_hash_u32(const void *key)
-{
-   return XXH32(key, 4, 0);
+   return _mesa_fnv32_1a_accumulate_block(_mesa_fnv32_1a_offset_bias,
+                                          data, size);
 }
 
 /** FNV-1a string hash implementation */
 uint32_t
 _mesa_hash_string(const void *_key)
 {
-   return _mesa_hash_string_with_length(_key, strlen((const char *)_key));
-}
-
-uint32_t
-_mesa_hash_string_with_length(const void *_key, unsigned length)
-{
-   uint32_t hash = 0;
+   uint32_t hash = _mesa_fnv32_1a_offset_bias;
    const char *key = _key;
-#if defined(_WIN64) || defined(__x86_64__)
-   hash = (uint32_t)XXH64(key, length, hash);
-#else
-   hash = XXH32(key, length, hash);
-#endif
+
+   while (*key != 0) {
+      hash = _mesa_fnv32_1a_accumulate(hash, *key);
+      key++;
+   }
+
    return hash;
-}
-
-uint32_t
-_mesa_hash_pointer(const void *pointer)
-{
-   uintptr_t num = (uintptr_t) pointer;
-   return (uint32_t) ((num >> 2) ^ (num >> 6) ^ (num >> 10) ^ (num >> 14));
-}
-
-bool
-_mesa_key_int_equal(const void *a, const void *b)
-{
-   return *((const int *)a) == *((const int *)b);
-}
-
-bool
-_mesa_key_uint_equal(const void *a, const void *b)
-{
-
-   return *((const unsigned *)a) == *((const unsigned *)b);
-}
-
-bool
-_mesa_key_u32_equal(const void *a, const void *b)
-{
-   return *((const uint32_t *)a) == *((const uint32_t *)b);
 }
 
 /**
@@ -733,21 +601,6 @@ _mesa_pointer_hash_table_create(void *mem_ctx)
 {
    return _mesa_hash_table_create(mem_ctx, _mesa_hash_pointer,
                                   _mesa_key_pointer_equal);
-}
-
-
-bool
-_mesa_hash_table_reserve(struct hash_table *ht, unsigned size)
-{
-   if (size < ht->max_entries)
-      return true;
-   for (unsigned i = ht->size_index + 1; i < ARRAY_SIZE(hash_sizes); i++) {
-      if (hash_sizes[i].max_entries >= size) {
-         _mesa_hash_table_rehash(ht, i);
-         break;
-      }
-   }
-   return ht->max_entries >= size;
 }
 
 /**
@@ -777,54 +630,22 @@ key_u64_equals(const void *a, const void *b)
 
 #define FREED_KEY_VALUE 0
 
-static void _mesa_hash_table_u64_delete_keys(void *data)
-{
-   struct hash_table_u64 *ht = ralloc_parent(data);
-
-   _mesa_hash_table_u64_clear(ht);
-}
-
 struct hash_table_u64 *
 _mesa_hash_table_u64_create(void *mem_ctx)
 {
    STATIC_ASSERT(FREED_KEY_VALUE != DELETED_KEY_VALUE);
    struct hash_table_u64 *ht;
 
-   ht = rzalloc(mem_ctx, struct hash_table_u64);
+   ht = CALLOC_STRUCT(hash_table_u64);
    if (!ht)
       return NULL;
 
    if (sizeof(void *) == 8) {
-      ht->table = _mesa_hash_table_create(ht, _mesa_hash_pointer,
+      ht->table = _mesa_hash_table_create(mem_ctx, _mesa_hash_pointer,
                                           _mesa_key_pointer_equal);
    } else {
-      ht->table = _mesa_hash_table_create(ht, key_u64_hash,
+      ht->table = _mesa_hash_table_create(mem_ctx, key_u64_hash,
                                           key_u64_equals);
-
-      /* Allocate a ralloc sub-context which takes the u64 hash table
-       * as a parent and attach a destructor to it so we can free the
-       * hash_key_u64 objects that were allocated by
-       * _mesa_hash_table_u64_insert().
-       *
-       * The order of creation of this sub-context is crucial: it needs
-       * to happen after the _mesa_hash_table_create() call to guarantee
-       * that the destructor is called before ht->table and its children
-       * are freed, otherwise the _mesa_hash_table_u64_clear() call in the
-       * destructor leads to a use-after-free situation.
-       */
-      if (ht->table) {
-         void *dummy_ctx = ralloc_context(ht);
-
-         /* If we can't allocate a sub-context, free the hash table
-          * immediately and return NULL to avoid future leaks.
-          */
-         if (!dummy_ctx) {
-            ralloc_free(ht);
-            return NULL;
-         }
-
-         ralloc_set_destructor(dummy_ctx, _mesa_hash_table_u64_delete_keys);
-      }
    }
 
    if (ht->table)
@@ -833,35 +654,66 @@ _mesa_hash_table_u64_create(void *mem_ctx)
    return ht;
 }
 
-static void
-_mesa_hash_table_u64_delete_key(struct hash_entry *entry)
-{
-   if (sizeof(void *) == 8)
-      return;
-
-   struct hash_key_u64 *_key = (struct hash_key_u64 *)entry->key;
-
-   if (_key)
-      FREE(_key);
-}
-
 void
-_mesa_hash_table_u64_clear(struct hash_table_u64 *ht)
+_mesa_hash_table_u64_clear(struct hash_table_u64 *ht,
+                           void (*delete_function)(struct hash_entry *entry))
 {
    if (!ht)
       return;
 
-   _mesa_hash_table_clear(ht->table, _mesa_hash_table_u64_delete_key);
-   ht->freed_key_data = NULL;
-   ht->deleted_key_data = NULL;
+   if (ht->deleted_key_data) {
+      if (delete_function) {
+         struct hash_table *table = ht->table;
+         struct hash_entry entry;
+
+         /* Create a fake entry for the delete function. */
+         if (sizeof(void *) == 8) {
+            entry.hash = table->key_hash_function(table->deleted_key);
+         } else {
+            struct hash_key_u64 _key = { .value = (uintptr_t)table->deleted_key };
+            entry.hash = table->key_hash_function(&_key);
+         }
+         entry.key = table->deleted_key;
+         entry.data = ht->deleted_key_data;
+
+         delete_function(&entry);
+      }
+      ht->deleted_key_data = NULL;
+   }
+
+   if (ht->freed_key_data) {
+      if (delete_function) {
+         struct hash_table *table = ht->table;
+         struct hash_entry entry;
+
+         /* Create a fake entry for the delete function. */
+         if (sizeof(void *) == 8) {
+            entry.hash = table->key_hash_function(uint_key(FREED_KEY_VALUE));
+         } else {
+            struct hash_key_u64 _key = { .value = (uintptr_t)FREED_KEY_VALUE };
+            entry.hash = table->key_hash_function(&_key);
+         }
+         entry.key = uint_key(FREED_KEY_VALUE);
+         entry.data = ht->freed_key_data;
+
+         delete_function(&entry);
+      }
+      ht->freed_key_data = NULL;
+   }
+
+   _mesa_hash_table_clear(ht->table, delete_function);
 }
 
 void
-_mesa_hash_table_u64_destroy(struct hash_table_u64 *ht)
+_mesa_hash_table_u64_destroy(struct hash_table_u64 *ht,
+                             void (*delete_function)(struct hash_entry *entry))
 {
    if (!ht)
       return;
-   ralloc_free(ht);
+
+   _mesa_hash_table_u64_clear(ht, delete_function);
+   _mesa_hash_table_destroy(ht->table, delete_function);
+   free(ht);
 }
 
 void
@@ -887,19 +739,7 @@ _mesa_hash_table_u64_insert(struct hash_table_u64 *ht, uint64_t key,
          return;
       _key->value = key;
 
-      struct hash_entry *entry =
-         hash_table_get_entry(ht->table, key_u64_hash(_key), _key);
-
-      if (!entry) {
-         FREE(_key);
-         return;
-      }
-
-      entry->data = data;
-      if (!entry_is_present(ht->table, entry))
-         entry->key = _key;
-      else
-         FREE(_key);
+      _mesa_hash_table_insert(ht->table, _key, data);
    }
 }
 
@@ -957,70 +797,6 @@ _mesa_hash_table_u64_remove(struct hash_table_u64 *ht, uint64_t key)
       struct hash_key *_key = (struct hash_key *)entry->key;
 
       _mesa_hash_table_remove(ht->table, entry);
-      FREE(_key);
-   }
-}
-
-
-/*
- * Iterates in order ("freed key", "deleted key", regular entries...)
- */
-struct hash_entry_u64
-_mesa_hash_table_u64_next_entry(struct hash_table_u64 *ht,
-                                struct hash_entry_u64 *ent)
-{
-   /* First entry: freed key */
-   if (!ent && ht->freed_key_data) {
-      return (struct hash_entry_u64){
-         .key = FREED_KEY_VALUE,
-         .data = ht->freed_key_data,
-      };
-   }
-
-   /* Second entry: deleted key */
-   if ((!ent || ent->key == FREED_KEY_VALUE) && ht->deleted_key_data) {
-      return (struct hash_entry_u64){
-         .key = DELETED_KEY_VALUE,
-         .data = ht->deleted_key_data,
-      };
-   }
-
-   /* All other entries: regular */
-   struct hash_entry *next =
-      _mesa_hash_table_next_entry(ht->table, ent ? ent->_entry : NULL);
-
-   if (!next)
-      return (struct hash_entry_u64){.data = NULL};
-
-   uint64_t key;
-   if (sizeof(void *) == 8) {
-      key = (uintptr_t)next->key;
-   } else {
-      const struct hash_key_u64 *_key = next->key;
-      key = _key->value;
-   }
-
-   return (struct hash_entry_u64){
-      .key = key,
-      .data = next->data,
-      ._entry = next,
-   };
-}
-
-/* Updates the data of a u64 hash_table entry inside a
- * hash_table_u64_foreach() loop
- */
-void
-_mesa_hash_table_u64_replace(struct hash_table_u64 *ht,
-                             const struct hash_entry_u64 *ent,
-                             void *new_data)
-{
-   if (ent->_entry) {
-      ent->_entry->data = new_data;
-   } else if (ent->key == FREED_KEY_VALUE) {
-      ht->freed_key_data = new_data;
-   } else {
-      assert(ent->key == DELETED_KEY_VALUE);
-      ht->deleted_key_data = new_data;
+      free(_key);
    }
 }
